src/main/webapp/app/entities/activation/delete/activation-delete-dialog.component.ts
import {Component} from '@angular/core';
import {FormsModule} from '@angular/forms';
import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import {IActivation} from '../activation.model';
import {ActivationService} from '../service/activation.service';
import {ITEM_DELETED_EVENT} from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './activation-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class ActivationDeleteDialogComponent {
  activation?: IActivation;

  constructor(protected activationService: ActivationService, protected activeModal: NgbActiveModal) {
  }

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.activationService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
